% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_select_kernel_params}
\alias{auto_select_kernel_params}
\title{Automatically Select Kernel Parameters}
\usage{
auto_select_kernel_params(kernel, Y)
}
\arguments{
\item{kernel}{Kernel function to use.}

\item{Y}{A numeric matrix of dimension (\code{n}, \code{p}), where each column corresponds to the observed trajectory of a variable. Rows align with \code{obs_time}.}
}
\value{
A list of length \code{p}, where each element is a named list of
parameters for a specific variable (e.g., \code{list(bandwidth = 1)} for
Gaussian kernel). If the list has length 1, the same parameter set is used
for all variables.
}
\description{
Currently supports a data-driven bandwidth selection for the Gaussian kernel,
with the following fixed settings for other kernels:
\itemize{
\item Linear kernel: no parameters.
\item Polynomial kernel: cubic polynomial (\code{intercept = 1}, \code{degree = 3}).
\item Matern kernel: length-scale fixed at 1.
}
}
\examples{
set.seed(1)
obs_time <- seq(0, 1, length.out = 10)
Y <- cbind(sin(2 * pi * obs_time), cos(4 * pi * obs_time)) + 0.1 * matrix(rnorm(20), 10, 2)  # each col is a variable
auto_select_kernel_params("gaussian", Y)

}
