% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune_network}
\alias{prune_network}
\title{Title}
\usage{
prune_network(
  network_original,
  prune_thres = 0.05,
  depth = NULL,
  eval_edge_R2_pruned = TRUE,
  Y_list,
  yy_smth_list,
  obs_time_list,
  tt,
  kernel,
  kernel_params_list,
  interaction_term,
  theta_initial_list = NULL,
  max_iter = 5,
  tol = 0.001,
  verbose = 0
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
