% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_kernel_ODE.R
\name{refit_kernel_ODE}
\alias{refit_kernel_ODE}
\title{Title}
\usage{
refit_kernel_ODE(
  Y,
  obs_time,
  yy_smth,
  tt,
  kernel,
  kernel_params,
  interaction_term,
  theta_initial = NULL,
  adj_matrix = NULL,
  nzero_thres = NULL,
  tol = 0.001,
  max_iter = 10,
  verbose = 0
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
