% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelODE_step2.R
\name{kernelODE_step2}
\alias{kernelODE_step2}
\title{Title}
\usage{
kernelODE_step2(
  Y,
  obs_time,
  yy_smth,
  tt,
  kernel,
  kernel_params,
  interaction_term = FALSE,
  theta_initial = NULL,
  adj_matrix = NULL,
  nzero_thres = NULL,
  eval_edge_R2 = FALSE,
  eval_loss = FALSE,
  tol = 0.001,
  max_iter = 10,
  parallel = TRUE,
  verbose = 0
)
}
\arguments{
\item{verbose}{}
}
\description{
This is an iterative process.
}
